ability to switch to previously uploaded avatar
Presently, if a user switches to gravatar after uploading their image,
they will not be able to switch back from the front-end.

This commit introduces a new issue, though. What happens if the user
doesn't already have an uploaded image?

The problem is that if the user is having "initials" or "gravatar", there
is no way for the frontend to know if an upload already exists.
Maybe once an API is built for that we can use that to check if this
option should be made visible or not.